+++ test/functional/test_download_and_import.py
@@ -12,11 +12,13 @@
 # language governing permissions and limitations under the License.
 from __future__ import absolute_import
 
+import importlib
 import shlex
 import subprocess
 import textwrap
 
 import pytest
+import six
 
 from sagemaker_containers.beta.framework import errors, modules
 import test
@@ -27,6 +29,15 @@ data = ['from distutils.core import setup\n',
 SETUP_FILE = test.File('setup.py', data)
 
 USER_SCRIPT_FILE = test.File('my_test_script.py', 'def validate(): return True')
+
+REQUIREMENTS_TXT_ASSERT_STR = """
+ ____                   __  __       _.............
+/ ___|  __ _  __ _  ___|  \/  | __ _| | _____ _ __.
+\___ \ / _` |/ _` |/ _ \ |\/| |/ _` | |/ / _ \ '__|
+ ___) | (_| | (_| |  __/ |  | | (_| |   <  __/ |...
+|____/ \__,_|\__, |\___|_|  |_|\__,_|_|\_\___|_|...
+             |___/.................................
+""".replace('.', ' ').strip()  # noqa W605
 
 
 @pytest.fixture(name='user_module_name')
@@ -74,6 +85,32 @@ def test_import_module_with_local_script(user_module, user_module_name, tmpdir):
     assert module.validate()
 
 
+@pytest.mark.parametrize('user_module',
+                         [test.UserModule(USER_SCRIPT_FILE).add_file(SETUP_FILE),
+                          test.UserModule(USER_SCRIPT_FILE)])
+def test_import_module_via_download_and_install(user_module, user_module_name):
+    user_module.upload()
+
+    modules.download_and_install(user_module.url, name=user_module_name, cache=False)
+    module = importlib.import_module(user_module_name)
+    six.moves.reload_module(module)
+
+    assert module.validate()
+
+
+@pytest.mark.parametrize('user_module',
+                         [test.UserModule(USER_SCRIPT_FILE).add_file(SETUP_FILE),
+                          test.UserModule(USER_SCRIPT_FILE)])
+def test_import_module_with_s3_script_via_download_and_install(user_module, user_module_name):
+    user_module.upload()
+
+    modules.download_and_install(user_module.url, name=user_module_name, cache=False)
+    module = importlib.import_module(user_module_name)
+    six.moves.reload_module(module)
+
+    assert module.validate()
+
+
 data = textwrap.dedent("""
             from pyfiglet import Figlet
 
@@ -95,14 +132,20 @@ def test_import_module_with_s3_script_with_requirements(user_module, user_module
 
     module = modules.import_module(user_module.url, user_module_name, cache=False)
 
-    assert module.say() == """
- ____                   __  __       _.............
-/ ___|  __ _  __ _  ___|  \/  | __ _| | _____ _ __.
-\___ \ / _` |/ _` |/ _ \ |\/| |/ _` | |/ / _ \ '__|
- ___) | (_| | (_| |  __/ |  | | (_| |   <  __/ |...
-|____/ \__,_|\__, |\___|_|  |_|\__,_|_|\_\___|_|...
-             |___/.................................
-""".replace('.', ' ').strip()  # noqa W605
+    assert module.say() == REQUIREMENTS_TXT_ASSERT_STR
+
+
+@pytest.mark.parametrize('user_module',
+                         [test.UserModule(USER_SCRIPT_WITH_REQUIREMENTS).add_file(SETUP_FILE),
+                          test.UserModule(USER_SCRIPT_WITH_REQUIREMENTS)])
+def test_import_module_with_requirements_via_download_and_install(user_module, user_module_name):
+    user_module = user_module.add_file(REQUIREMENTS_FILE).upload()
+
+    modules.download_and_install(user_module.url, name=user_module_name, cache=False)
+    module = importlib.import_module(user_module_name)
+    six.moves.reload_module(module)
+
+    assert module.say() == REQUIREMENTS_TXT_ASSERT_STR
 
 
 data = ['raise ValueError("this script does not work")']